move check permission to middleware
