+++ awsServer/app.js
@@ -3,9 +3,7 @@
 const express = require('express');
 const helmet = require('helmet');
 const awesomeLogger = require('express-bunyan-logger');
-const auth = require('basic-auth');
 
-const { credentials } = require('../simpleAuth.json');
 const { getList, deleteItem, moveItem, giveSignedUrl } = require('./awsFunctions');
 
 const possibleFolder = ['richard', 'car', 'people'];
@@ -17,18 +15,6 @@ app.use((req, res, next) => {
   res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
   next();
 });
-
-// Way too simple for production but ok for Home fun
-const checkPermission = (req, res, next) => {
-  const userInput = auth(req);
-  if (!userInput || !credentials[userInput.name] || userInput.pass !== credentials[userInput.name]) {
-    res.statusCode = 401;
-    res.setHeader('WWW-Authenticate', 'Basic realm="example"');
-    res.end('Access denied');
-  } else {
-    next();
-  }
-};
 
 app.use(awesomeLogger({
   name: 'logger',
